move _version_tools into their own place and leave _version somewhat isolated
